raspberry/requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
Mock.GPIO==0.1.9
netifaces==0.11.0
packaging==24.2
paho-mqtt==2.1.0
pluggy==1.5.0
pytest==8.3.4
pytest-mock==3.14.0
tomli==2.2.1
